+++ test/power_geojson_test.dart
@@ -1,7 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-class MockPowerGeojsonPlatform with MockPlatformInterfaceMixin {
+class MockPowerGeojsonPlatform {
   Future<String?> getPlatformVersion() => Future.value('42');
 }